Update: project architecture and artist binaries.
