+++ Source/BCR/Sources/System/AWood.cpp
@@ -1,0 +1,27 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "BCR/Headers/System/AWood.h"
+
+// Sets default values
+AAWood::AAWood()
+{
+ 	// Set this actor to call Tick() every frame.  You can turn this off to improve performance if you don't need it.
+	PrimaryActorTick.bCanEverTick = true;
+
+}
+
+// Called when the game starts or when spawned
+void AAWood::BeginPlay()
+{
+	Super::BeginPlay();
+	
+}
+
+// Called every frame
+void AAWood::Tick(float DeltaTime)
+{
+	Super::Tick(DeltaTime);
+
+}
+
